Various improvements

+++ sdk-android/src/main/java/cu/suitetecsa/sdk/ussd/model/BonusData.kt
@@ -1,5 +1,5 @@
 package cu.suitetecsa.sdk.ussd.model
 
 data class BonusData(
-    val bonusDataCount: Long?, val bonusDataCountLte: Long?, val bonusDataDueDate: String?
+    val bonusDataCount: Double?, val bonusDataCountLte: Double?, val bonusDataDueDate: String?
 )
